+++ connector_dns/connector.py
@@ -19,7 +19,7 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-from openerp.osv import orm, fields
+from openerp.osv import models, orm, fields
 from openerp.addons.connector.connector import (Environment,
                                                 install_in_connector)
 from openerp.addons.connector.checkpoint import checkpoint
@@ -33,7 +33,7 @@ class DNSConnectorInstalled(orm.AbstractModel):
 
     If the model is in the registry, the module is installed.
     """
-    _name = 'dns_connector.installed'
+    _name = 'connector_dns.installed'
 
 
 def get_environment(session, model_name, backend_id):
@@ -43,7 +43,7 @@ def get_environment(session, model_name, backend_id):
     return env
 
 
-class DNSBinding(orm.AbstractModel):
+class DNSBinding(models.Model):
     """ Abstract Model for the Bindigs.
     All the models used as bindings between dnspod and OpenERP
     (``dnspod.res.partner``, ``dnspod.product.product``, ...) should
@@ -53,22 +53,20 @@ class DNSBinding(orm.AbstractModel):
     _inherit = 'external.binding'
     _description = 'dns Binding (abstract)'
 
-    _columns = {
-        # 'openerp_id': openerp-side id must be declared in concrete model
-        'backend_id': fields.many2one(
-            'dns.backend',
-            'DNS Backend',
-            required=True,
-            ondelete='restrict'),
-        # fields.char because 0 is a valid dnspod ID
-        'dns_id': fields.char('ID on other software'),
-        # state of the record synchronization with dnspod
-        'state': fields.selection(
-            [('draft', 'Draft'), ('done', 'Done'),
-             ('exception', 'Exception')], 'State',
-            default="draft",
-            help='Done when succeed otherwise Exception'),
-    }
+    # 'openerp_id': openerp-side id must be declared in concrete model
+    backend_id = fields.Many2one(
+        'dns.backend',
+        String='DNS Backend',
+        required=True,
+        ondelete='restrict'),
+    # fields.char because 0 is a valid dnspod ID
+    dns_id = fields.Char('ID on other software'),
+    # state of the record synchronization with dnspod
+    state = fields.Selection(
+        [('draft', 'Draft'), ('done', 'Done'),
+         ('exception', 'Exception')], 'State',
+        default="draft",
+        help='Done when succeed otherwise Exception'),
 
 
 def add_checkpoint(session, model_name, record_id, backend_id):
