+++ connector_dns/dns.py
@@ -27,7 +27,7 @@ class DNSBackend(models.Model):
     _inherit = 'connector.backend'
     _backend_type = 'dns'
 
-    def select_version(self):
+    def _select_version(self):
         return []
 
     login = fields.Char('Login', help='Provider\'s login.', required=True)
@@ -38,7 +38,8 @@ class DNSBackend(models.Model):
          ('exception', 'Exception')], 'State', default="draft",
         help='"Confirmed" when the domain has been succesfully created.')
     version = fields.Selection(
-        select_version, string='Service Provider', help='DNS service provider',
+        '_select_version', string='Service Provider',
+        help='DNS service provider',
         required=True)
 
     @api.multi
@@ -63,11 +64,10 @@ class DNSRecord(models.Model):
     _name = 'dns.record'
     _inherit = 'dns.binding'
 
-    def line_select_version(self):
-        res = []
-        return res
+    def _line_select_version(self):
+        return []
 
-    def type_select_version(self):
+    def _type_select_version(self):
         return []
 
     name = fields.Char('Sub domain', help="host record,such as 'www'",
@@ -75,8 +75,8 @@ class DNSRecord(models.Model):
     domain_id = fields.Many2one(
         'dns.domain', string="Domain", domain="[('state','=','done')]",
         help="Domain which has already confirmed")
-    type = fields.Selection(type_select_version, string='Record Type')
-    line = fields.Selection(line_select_version, string='Record Line')
+    type = fields.Selection('type_select_version', string='Record Type')
+    line = fields.Selection('line_select_version', string='Record Line')
     value = fields.Text('Value', help="such as IP:200.200.200.200",
                         required=True)
     mx_priority = fields.Integer(string='MX priority', help="scope:1-20",
