+++ connector_dns_dnspod/readme.rst
@@ -1,0 +1,63 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+    :alt: License
+
+Kit Product Inventory
+=====================
+
+This module aims to add more flexibility on product inventory.
+  * Allowing user to add kit products and manage those kit products inventory.
+
+  * Allowing user to know the quantity on hand of kit products.
+
+Please note that this module re-defines the create and write method of mrp.bom.
+
+Installation
+============
+
+To install this module, you need to:
+
+ * have basic modules installed (stock, product, mrp, sale)
+
+Configuration
+=============
+
+To configure this module, you need to:
+
+ * No specific configuration needed.
+
+Usage
+=====
+
+
+For further information, please visit:
+
+ * https://www.odoo.com/forum/help-1
+
+Known issues / Roadmap
+======================
+
+
+Credits
+=======
+
+
+Contributors
+------------
+
+* Alex Duan <alex.duan@elico-corp.com>
+* Liu Lixia <liu.lixia@elico-corp.com>
+
+Maintainer
+----------
+
+.. image:: http://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: http://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization
+    whose mission is to support the collaborative development of Odoo features
+        and promote its widespread use.
+
+To contribute to this module, please visit http://odoo-community.org.
